User registration and authentication done

+++ mysite/simplefeed/views.py
@@ -7,7 +7,8 @@ import urllib, re
 from bs4 import BeautifulSoup
 import difflib
 from lxml.html.diff import htmldiff
-
+from django.contrib.auth.models import Group, User
+from django.contrib.auth import authenticate, login, logout
 def diff(h1, h2):
     print h1
     print h2
@@ -47,6 +48,8 @@ def parsehtml(url):
     return s
 
 def index(request):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect('/feed/login_show')
     target_list = Target.objects.order_by('url').all()
     for t in target_list:
         curr = gethtml(t.url)
@@ -62,8 +65,47 @@ def index(request):
     }
     return HttpResponse(template.render(context, request))
 
+def register_show(request):
+    template = loader.get_template('simplefeed/register.html')
+    context = {
+    }
+    return HttpResponse(template.render(context, request))
+
+def register(request):
+    # Add user already exists
+    User.objects.create_user(username = request.POST['username'], password = request.POST['password'])
+    return HttpResponseRedirect('/feed/')
+
+def login_show(request):
+    template = loader.get_template('simplefeed/login.html')
+    context = {
+    }
+    return HttpResponse(template.render(context, request))
+
+def loginuser(request):
+    username = request.POST['username']
+    password = request.POST['password']
+    user = authenticate(username=username, password=password)
+    if user is not None:
+        if user.is_active:
+            login(request, user)
+            return HttpResponseRedirect('/feed')
+            # Redirect to a success page.
+        else:
+            return HttpResponse("Disabled account")
+            # Return a 'disabled account' error message
+
+    else:
+        # Return an 'invalid login' error message.
+        return HttpResponseRedirect('/feed/login_show')
+
+def logoutuser(request):
+    logout(request)
+    return HttpResponseRedirect('/feed/')
 
 def addform(request):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect('/feed/login_show')
     template = loader.get_template('simplefeed/addform.html')
     context = {
     }
@@ -71,12 +113,15 @@ def addform(request):
 
 
 def add(request):
-
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect('/feed/login_show')
     q = Target(url = request.POST['url'], category = request.POST['category'], content = parsehtml(request.POST['url']))
     q.save()
     return HttpResponseRedirect('/feed/')
 
 def updateform(request, target_id):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect('/feed/login_show')
     template = loader.get_template('simplefeed/updateform.html')
     q = Target.objects.get(pk = target_id)
     context = {
@@ -85,6 +130,8 @@ def updateform(request, target_id):
     return HttpResponse(template.render(context, request))
 
 def update(request, target_id):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect('/feed/login_show')
     q = Target.objects.get(pk = target_id)
     q.url = request.POST['url']
     q.category = request.POST['category']
@@ -94,6 +141,8 @@ def update(request, target_id):
     return HttpResponseRedirect("/feed/")
 
 def delete(request, target_id):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect('/feed/login_show')
     q = Target.objects.get(pk = target_id)
     q.delete()
     return HttpResponseRedirect("/feed/")
